+++ packages/types/global.d.ts
@@ -1,0 +1,6 @@
+import { SnapProvider } from './src';
+
+// Types that should be available globally within a Snap
+declare global {
+  const wallet: SnapProvider;
+}